pyproject.toml
[project]
name = "realtime-mvp"
version = "0.1.0"
description = "Real-time marketing analytics MVP with Kafka, Flink, and Streamlit"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "kafka-python>=2.0.2",
    "pandas>=2.0.0",
    "faker>=20.0.0",
    "numpy>=1.24.0",
    "apache-flink>=1.17.0",
    "streamlit>=1.28.0",
    "requests>=2.31.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
